--- src/main/java/dev/book/user/UserController.java
@@ -1,27 +1,0 @@
-package dev.book.user;
-
-import dev.book.global.config.security.dto.CustomUserDetails;
-import dev.book.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/users")
-public class UserController {
-
-    private final UserService userService;
-
-    @GetMapping("/profile")
-    public ResponseEntity<?> getUserProfile(@AuthenticationPrincipal CustomUserDetails userDetails){
-        return ResponseEntity.ok()
-                .body(userService.getUserProfile(userDetails));
-    }
-
-    //todo 업적, 통계, 경고 반환
-
-}
