+++ src/main/java/dev/book/user/dto/request/UserProfileUpdateRequest.java
@@ -1,0 +1,4 @@
+package dev.book.user.dto.request;
+
+public record UserProfileUpdateRequest(String nickname, String profileImageUrl) {
+}
